iqueryable, thenby, refactoring
