+++ src/NetSort.UnitTests/EF/EF_Person.cs
@@ -1,0 +1,8 @@
+﻿namespace NetSort.UnitTests.EF
+{
+    public class EF_Person
+    {
+        [Sortable("age")]
+        public int Age { get; set; }
+    }
+}
