+++ src/NetSort.UnitTests/EF/FakeDbContext.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.EntityFrameworkCore;
+
+namespace NetSort.UnitTests.EF
+{
+    public class FakeDbContext : DbContext
+    {
+        public virtual DbSet<EF_Person> EF_Persons { get; set; }
+    }
+}
