+++ src/NetSort.UnitTests/SortableExtensionsTests.cs
@@ -4,7 +4,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.Text;
 
 namespace NetSort.UnitTests
 {
