+++ src/NetSort.UnitTests/ThenBySortableExtensionsTests.cs
@@ -1,10 +1,8 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using NetSort.UnitTests.Models;
-using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.Text;
 
 namespace NetSort.UnitTests
 {
