+++ src/NetSort/IQueryableThenBySortableExtensions.cs
@@ -1,0 +1,107 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace NetSort
+{
+    public static partial class SortableExtensions
+    {
+        /// <summary>
+        /// Returns an ordered queryable per the specified 'key' in the order specified on the 'SortableAttribute' of the key passed in, or the default (Ascending).
+        /// </summary>
+        /// <typeparam name="T">The type of data being sorted</typeparam>
+        /// <param name="items">The data to sort</param>
+        /// <param name="key">The 'key' to sort by. This key must be present as the 'SortKey' of exactly one 'SortableAttribute' on a property of type T</param>
+        /// <param name="delimiter">The char used to separate nested properties</param>
+        public static IOrderedQueryable<T> ThenSortByKey<T>(this IOrderedQueryable<T> items, string key, char delimiter) where T : class
+        {
+            return DoThenSort(items, key, delimiter, null);
+        }
+
+        /// <summary>
+        /// Returns an ordered queryable per the specified 'key' and 'order'.
+        /// </summary>
+        /// <typeparam name="T">The type of data being sorted</typeparam>
+        /// <param name="items">The data to sort</param>
+        /// <param name="key">The 'key' to sort by. This key must be present as the 'SortKey' of exactly one 'SortableAttribute' on a property of type T</param>
+        /// <param name="direction">The direction to sort (Asc | Desc)</param>
+        /// <param name="delimiter">The char used to separate nested properties</param>
+        public static IOrderedQueryable<T> ThenSortByKey<T>(this IOrderedQueryable<T> items, string key, char delimiter, SortDirection direction) where T : class
+        {
+            return DoThenSort(items, key, delimiter, direction);
+        }
+
+        /// <summary>
+        /// Returns an ordered queryable per the specified 'key' in the order specified on the 'SortableAttribute' of the key passed in, or the default (Ascending).
+        /// </summary>
+        /// <typeparam name="T">The type of data being sorted</typeparam>
+        /// <param name="items">The data to sort</param>
+        /// <param name="key">The 'key' to sort by. This key must be present as the 'SortKey' of exactly one 'SortableAttribute' on a property of type T</param>
+        public static IOrderedQueryable<T> ThenSortByKey<T>(this IOrderedQueryable<T> items, string key) where T : class
+        {
+            return ThenSortByKey(items, key, Constants.DEFAULT_NESTED_PROP_DELIMITER);
+        }
+
+        /// <summary>
+        /// Returns an ordered queryable per the specified 'key' and 'order'.
+        /// </summary>
+        /// <typeparam name="T">The type of data being sorted</typeparam>
+        /// <param name="items">The data to sort</param>
+        /// <param name="key">The 'key' to sort by. This key must be present as the 'SortKey' of exactly one 'SortableAttribute' on a property of type T</param>
+        /// <param name="direction">The direction to sort (Asc | Desc)</param>
+        public static IOrderedQueryable<T> ThenSortByKey<T>(this IOrderedQueryable<T> items, string key, SortDirection direction) where T : class
+        {
+            return ThenSortByKey(items, key, Constants.DEFAULT_NESTED_PROP_DELIMITER, direction);
+        }
+
+        /// <summary>
+        /// Returns an ordered queryable per the specified 'key' in the specified 'order'.
+        /// </summary>
+        /// <typeparam name="T">The type of data being sorted</typeparam>
+        /// <param name="items">The data to sort</param>
+        /// <param name="key">The 'key' to sort by. This key must be present as the 'SortKey' of exactly one 'SortableAttribute' on a property of type T</param>
+        /// <param name="direction">The direction to sort (Asc | Desc)</param>
+        /// <param name="delimiter">The char used to separate nested properties</param>
+        public static IOrderedQueryable<T> ThenSortByKey<T>(this IOrderedQueryable<T> items, string key, char delimiter, string direction) where T : class
+        {
+            var directionEnum = ParseDirection(direction);
+            return ThenSortByKey(items, key, delimiter, directionEnum);
+        }
+
+        /// <summary>
+        /// Returns an ordered queryable per the specified 'key' in the specified 'order'.
+        /// </summary>
+        /// <typeparam name="T">The type of data being sorted</typeparam>
+        /// <param name="items">The data to sort</param>
+        /// <param name="key">The 'key' to sort by. This key must be present as the 'SortKey' of exactly one 'SortableAttribute' on a property of type T</param>
+        /// <param name="direction">The direction to sort (Asc | Desc)</param>
+        public static IOrderedQueryable<T> ThenSortByKey<T>(this IOrderedQueryable<T> items, string key, string direction) where T : class
+        {
+            return ThenSortByKey(items, key, Constants.DEFAULT_NESTED_PROP_DELIMITER, direction);
+        }
+
+        private static IOrderedQueryable<T> DoThenSort<T>(IOrderedQueryable<T> items, string key, char delimiter, SortDirection? direction) where T : class
+        {
+            var metadata = SortOperationMetadataFinder.Find<T>(key, delimiter, direction);
+            if (!metadata.Any())
+            {
+                var error = $"A 'sortable attribute' could not be found.\n Key: {key} \nType: {typeof(T).Name}.";
+                throw new ArgumentException(error);
+            }
+
+            return ThenSort(items, metadata);
+        }
+
+        private static IOrderedQueryable<T> ThenSort<T>(IOrderedQueryable<T> items, IEnumerable<SortOperationMetadata> metadata)
+        {
+            if (metadata.Last().Direction == SortDirection.Asc)
+            {
+                return items.ThenBy(i => GetNestedValue(metadata, i));
+            }
+            else
+            {
+                return items.ThenByDescending(i => GetNestedValue(metadata, i));
+            }
+        }
+    }
+}
