+++ src/NetSort/SortableAttribute.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+
 namespace NetSort
 {
     /// <summary>
