+++ src/NetSort/Validation/SortKeyValidator.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
 
 namespace NetSort.Validation
 {
